backend/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.1

# ── Authentication ─────────────────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ── CORS (allow React frontend to call the API) ────────────────────────────────
django-cors-headers==4.3.1

# ── Filtering (filter transactions by month, category, type…) ─────────────────
django-filter==24.2

# ── Environment variables (.env file) ─────────────────────────────────────────
python-decouple==3.8

# ── Database ───────────────────────────────────────────────────────────────────
# SQLite is built into Python — no extra package needed for dev.
# Uncomment the line below when deploying to production with PostgreSQL:
# psycopg2-binary==2.9.9

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-django==4.8.0
factory-boy==3.3.0
